feat(helm): rework gateway probes definition with default value and customisation
This commit will keep the existing startup/liveness/readiness probe in
values.yaml but change the current behaviour:

Now we define default values which are overwrite when customer define
something in values.

It is backward compatible to ensure that existing values continue to
work as previously.

Also I asked for a deprecation of `gateway.readinessProbe.domainSync` as
we can overwrite it in a "better" way now.

TT-5491

[skip ci]