feat(helm): rework api probes definition with default value and customisation
This commit will keep the existing startup/liveness/readiness probe in
values.yaml but change the current behaviour:

Now we define default values which can be overwrite when customer define
something in values.

It is backward compatible to ensure that existing values continue to
work as previously.

Also, all probes are compute in dedicated template.

The scope of this commit is only about api probes.

TT-5491